% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_scalebar}
\alias{add_scalebar}
\title{Add scale bar to spatial plot}
\usage{
add_scalebar(p, seurat_obj, bar_length_um = 1000)
}
\arguments{
\item{p}{A ggplot object}

\item{seurat_obj}{The Seurat object used to create the plot}

\item{bar_length_um}{Scale bar length in micrometers (default: 1000)}
}
\value{
The plot with scale bar added
}
\description{
Adds a scale bar to a ggplot spatial plot based on Visium spot spacing.
}
\examples{
\dontrun{
p <- plot_spatial_feature(seurat_obj, "COL1A1")
p <- add_scalebar(p, seurat_obj)
}

}
