% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential-expression.R
\name{differential-expression}
\alias{differential-expression}
\title{ECM Differential Expression Analysis}
\description{
Functions for differential expression analysis of ECM/matrisome genes
in spatial transcriptomics data.
}
