% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm-annotation.R
\name{ecm-annotation}
\alias{ecm-annotation}
\title{ECM Niche Annotation Functions}
\description{
Functions for classifying spots into ECM niches using the
ScType algorithm.

ScType functions adapted from:
Ianevski et al., 2022 (https://github.com/IanevskiAleksandr/sc-type)
Licensed under GPL-3.0
}
