% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ecm_markers}
\alias{ecm_markers}
\title{ECM Niche Markers}
\format{
A list with two elements:
\describe{
\item{gs_positive}{Positive marker genes per niche}
\item{gs_negative}{Negative marker genes per niche}
}
}
\usage{
ecm_markers
}
\description{
Marker genes for ECM niche classification using ScType.
}
\examples{
data(ecm_markers)
names(ecm_markers$gs_positive)

}
\keyword{datasets}
