% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ecm_signatures}
\alias{ecm_signatures}
\title{ECM Niche Signatures}
\format{
A named list with three elements:
\describe{
\item{Vascular}{Genes associated with vascular ECM}
\item{Basement}{Genes associated with basement membrane}
\item{Interstitial}{Genes associated with interstitial matrix}
}
}
\usage{
ecm_signatures
}
\description{
Gene signatures for three ECM niche types used for UCell scoring.
}
\examples{
data(ecm_signatures)
names(ecm_signatures)

}
\keyword{datasets}
