% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrisome-pairs.R
\name{matrisome-pairs}
\alias{matrisome-pairs}
\title{Matrisome Pair Co-expression Functions}
\description{
Functions for spatial co-expression analysis of MatriComDB
pairs in which at least one member is a matrisome gene.
}
