% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{palette_ecm}
\alias{palette_ecm}
\title{Generate ECM niche color palette}
\usage{
palette_ecm(levels_vector)
}
\arguments{
\item{levels_vector}{Character vector of ECM niche levels}
}
\value{
Named character vector of hex colors
}
\description{
Creates a color palette for ECM niche annotations with biologically
meaningful colors: Vascular (red), Basement (purple), Interstitial (green).
}
\details{
ECM Niche Color Palette:
\itemize{
\item Vascular (#d42626): Red - blood vessels and vascular ECM
\item Basement (#6a3d9a): Purple - structural foundation
\item Interstitial (#2b9e2b): Green - matrix between cells
\item not.assigned (#7f7f7f): Grey - unassigned spots
}
}
\examples{
\dontrun{
colors <- palette_ecm(c("Vascular ECM", "Basement ECM", "Interstitial ECM"))
}

}
