% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepare_object}
\alias{prepare_object}
\title{Prepare Seurat object for matrisome analysis}
\usage{
prepare_object(
  seurat_obj,
  mat_feat_sigs = NULL,
  ecm_ucell_sigs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{mat_feat_sigs}{List of matrisome gene signatures (defaults to package data)}

\item{ecm_ucell_sigs}{List of ECM niche signatures (defaults to package data)}

\item{verbose}{Print progress messages}
}
\value{
Preprocessed Seurat object
}
\description{
Master preprocessing function that checks and adds required components:
\enumerate{
\item SCTransform (if needed for ScType)
\item Matrisome feature scores (UCell)
\item ECM niche classification (ScType)
\item ECM niche signatures
}
}
\examples{
\dontrun{
seurat_obj <- prepare_object(seurat_obj)
}

}
