% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial}
\alias{spatial}
\title{Spatial Statistics Functions}
\description{
Functions for computing spatial statistics including LISA
(Local Indicators of Spatial Association) and Moran's I autocorrelation.
}
