% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{translate_signatures}
\alias{translate_signatures}
\title{Translate gene signatures using HGNChelper}
\usage{
translate_signatures(seurat_obj, signature_list)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{signature_list}{Named list of gene vectors}
}
\value{
Named list of corrected gene vectors
}
\description{
Checks gene symbols and corrects outdated/alias symbols to match
what's present in the Seurat object.
}
\examples{
\dontrun{
sigs <- translate_signatures(seurat_obj, my_signatures)
}

}
