% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{lr_database}
\alias{lr_database}
\title{Ligand-Receptor Interaction Database}
\format{
A data frame with columns:
\describe{
\item{Ligand}{Ligand gene symbol}
\item{Receptor}{Receptor gene symbol}
\item{Source}{Database source for the interaction}
\item{Interaction_Type}{Type of interaction}
\item{Database_Score}{Confidence score from source database}
}
}
\usage{
lr_database
}
\description{
Database of ECM-related ligand-receptor interactions for spatial
co-expression analysis.
}
\details{
Contains 24,000+ deduplicated interactions curated from multiple databases,
with focus on ECM-related signaling.
}
\examples{
data(lr_database)
head(lr_database)
nrow(lr_database)

}
\keyword{datasets}
