% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrispace-package.R
\docType{package}
\name{matrispace-package}
\alias{matrispace-package}
\alias{matrispace}
\title{matrispace: Identification and Visualization of Spatially-Resolved ECM Gene Expression Patterns}
\description{
\pkg{matrispace} is the R package companion to MatriSpace. It brings the
same extracellular matrix-focused workflow used in the web application to
local, scriptable analysis of spatial transcriptomics data. The package
includes functions for:

\itemize{
\item Matrisome scoring using UCell signatures
\item ECM niche classification using the ScType algorithm
\item Spatial statistics (LISA, Moran's I)
\item Ligand-receptor co-expression analysis
\item Static visualizations for spatial data
}
}
\section{Main Analysis Functions}{

\describe{
\item{\code{\link{prepare_object}}}{Master preprocessing pipeline}
\item{\code{\link{score_matrisome}}}{Calculate matrisome category scores}
\item{\code{\link{annotate_ecm_niches}}}{Classify ECM niches with ScType}
\item{\code{\link{compute_lisa}}}{Local spatial association statistics}
\item{\code{\link{compute_morans_i}}}{Global spatial autocorrelation}
\item{\code{\link{score_lr_activity}}}{Ligand-receptor activity scoring}
}
}

\section{Data Objects}{

\describe{
\item{\code{\link{matrisome_db}}}{Gene-to-category mapping database}
\item{\code{\link{matrisome_signatures}}}{UCell gene signatures}
\item{\code{\link{ecm_signatures}}}{ECM niche signatures}
\item{\code{\link{lr_database}}}{Ligand-receptor interaction pairs}
}
}

\section{Visualization Functions}{

\describe{
\item{\code{\link{plot_spatial_feature}}}{Single feature spatial plot}
\item{\code{\link{plot_matrisome}}}{Matrisome category plots}
\item{\code{\link{plot_lisa}}}{LISA clustering visualization}
\item{\code{\link{plot_spatial_blend}}}{Two-feature blend plot}
}
}

\references{
\url{https://github.com/Theayomideo/matrispace}
}
\seealso{
The companion MatriSpace web application:
\url{http://matrinet.shinyapps.io/matrispace}
}
\author{
Ayomide Oshinjo, Daiqing Chen, Petar Petrov, Valerio Izzi, Alexandra Naba
}
